Increase height for items in shortcuts window.
This increases the height for each item in the keyboard shortcuts window from 10px to 12px.
While this may not seem like much, it ensures the text no longer overlaps for Japanese and Korean.
At also has the added benefit of making the window look a bit nicer for Western languages.