Fix scenery group assignment race conditions
When adding entries to scenery groups by hand, like we do for a lot of ToonTowner's stuff, it currently tries to get the loaded scenery group. If this happens not to be loaded at that time, this will fail.
Instead, hardcode the headers - that's how object .DATs do this anyway.