Refactor bottom toolbar: hex to dec, apply macro.
This was one of the few windows that had its widget positions defined in
hexadecimals rather than decimals.

The logic was further obscured by using a macro to offset an experimental
change to accomodate longer dates. This change has been widely embraced,
so this patch applied said macro to improve legibility.