Set DPI scaling on first launch
When OpenRCT2 is first launched every setting gets defaulted.
DPI default is set to 1.0 which is to small for high dpi monitors.
It's not a big issue as a scaling is already supported but hitting
the tremendous small upwards icon is hard. After this patch
on first launch the monitor scaling is checked and set to a more
appropriate value.

Signed-off-by: Tobias Kohlbau <tobias@kohlbau.de>