Fix #6958: Recent messages are not coloured
This fix is a workaround to what I believe is a bug in MSVC's optimised code generator. Somehow the casting of a wchar_t to a sint32 is causing some strange behaviour where the number 147 gets changed to 0x??93. Adding a printf to display codepoint was enough to alter the behaviour, making it very difficult to understand what was going on.