Fix memset byte value too big
`memset` takes an integer as argument which gets converted to an unsigned byte. `SPRITE_INDEX_NULL` has a value of 0xFFFF, essentially cutting it to 0xFF. Calling `std::fill_n` instead of `memcpy` fills the memory of each element rather than each byte. In this case, the result is the same.

None of the other calls to `memset` in the project use values bigger than 0xFF.