Fix #7569: Assertion hit when using the XY spinners
The X/Y spinners changed the current tile selected, and reset the element selection with it. Previously this would set the page to its default immediately, but since that's now done on invalidation instead, the page would not always be set. This is fixed by checking the index of the selected element too.