Fix #7601: Always-researched items not rendered with inset. (#7638)
This re-introduces the use of buffers to work around colours not being set properly.

For clipping strings, we also now manually invoke `gfx_clip_string` instead of relying
on `gfx_draw_string_left_clipped` to do it for us. The latter resets the current font,
meaning it prevents the use of `MEDIUM_EXTRA_DARK` bases. This issue should be addressed
in a later commit.