Fix #8151: Game freezes upon demolishing mazes at odd heights
Split MazeSetTrackAction.Query() into two functions
(QueryBuild and QueryOther).

Most of the code in Query() could only get executed if a new track
was being built. In other cases, checks were executed in every case,
even though they only made sense when building a new track on a new tile.

This issue #8151 since it was due to a check for height
that happened when demolishing the maze, even though
that check only made sense when building.

The check in question ---> "if ((_z & 0xF) != 0)"