Fix #8145: Mechanics fix vehicles not in station
If there was a pending breakdown for a vehicle during inspections,
mechanics would walk to that vehicle even if it was currently in the
middle of the track and not inside the station.

Instead, keep the mechanic in the "inspection" state and simply remove
the pending breakdown when the mechanic leaves the station after the
inspection. Also, when calling a mechanic, they will only be called to
fix the ride if the breakdown has already happened and is no longer
pending. If the breakdown is pending, the mechanic is only called for an
inspection.

Observing players might notice the mechanic will do a call before
leaving the station after the inspection if a pending breakdown was
fixed, and skip that call if there was no pending breakdown.

The staff stats will update to increase both, the number of inspections
as well as the number of fixed rides, in cases like this where a pending
breakdown is fixed during an inspection.