Android: target SDK 28, remove absolute path (#10568)
Android version: an absolute path was replaced with a call to the Environment class. The target SDK of 29 was decremented to a target SDK of 28 to fix an incompatibility with an SDK change. Game now builds; however a desktop version must be built first so that build artifacts are present in /bin.