Improvement: Option for Guests to use more of wide path
-Instead of only one wide flag, each path can have multiple levels of flags

-The first level runs the basic algorithm 8 times (or "groups") in
multiple directions for peeps to use paths symmetrically.

-Levels 2 and above each has 8 groups coresponding to the previous flags
to fill in the remaining spaces (Still incomplete)

-When peeps are path finding, the flag they consider when checking if a
path is wide is based on their ID. The original wide flag still keeps
track of the default update order. Staff peeps use the default wide
Flag.

-Each group and level of flags can be viewed with the debugger.

-Flags are stored in a array of vectors separate from the
tile structures

-In place of saving all flags, when loading a map the game will instead
load all wide flags when configured to use 1 or more levels. This means
the game state is not perfectly preserved upon saving when this feature
is on.

Future Improvements:
-Improve algorithm for widepath levels 2 and up. Some gaps are not yet covered.