Remove NEW_BLOCK_BRAKES define (#11727)
This define was never enabled, because this "new block brake" behaviour (i.e. not braking at all if the brake was open) was too different from the old behaviour to be useful.
Its intended purpose will be handled by adding actual speed settings on block brakes in the future.

This commit removes the define, and also removes the second parameter to apply_non_stop_block_brake(), which is now always true.