Fix Text Input window colours
Text Input uses one colour, specifically the body colour of the calling window. This was previously handles by a modified window shim, but https://github.com/OpenRCT2/OpenRCT2/pull/11653 changed this and this window now uses the standard window shim. This is why it now needs to change the colour assignment.