Restore spinner decrease button width to 13px
For the original spinners, I had made the decrease button 13px and the increase button 12px wide.
This was to ensure the minus glyph was properly aligned to the centre. There was an oversight
when I refactored these the other day, making both buttons 12px wide, and causing the glyph to
no longer be aligned properly.