Fix #12434, ENTERTAINER_COSTUME to strong enum.
Update functions to pass EntertainerCostume or arrays of
EntertainerCostume when possible instead of bare uint8_t.

There's still some places (when converting to sprites or back) where
there's a magic number 4 added or subtracted. But I'd rather not expand
the scope of this issue to mess with anything sprites if possible.